refactor: extract GitHub class
