+++ src/main/kotlin/ResultProcessor.kt
@@ -1,12 +1,5 @@
-import org.kohsuke.github.GHWorkflow
-import org.kohsuke.github.GHWorkflowRun
-import org.kohsuke.github.GitHubBuilder
-import java.io.FileOutputStream
 import java.nio.file.Path
-import java.util.zip.ZipEntry
-import java.util.zip.ZipInputStream
 import kotlin.io.path.absolutePathString
-import kotlin.io.path.createDirectories
 import kotlin.io.path.exists
 import kotlin.system.exitProcess
 
@@ -31,7 +24,8 @@ class ResultProcessor {
             exitProcess(1)
         }
 
-        downloadResults()
+        GitHub.downloadPreviousArtifact(Git.baseBranch, baselineResultsDir, artifactName)
+        GitHub.downloadPreviousArtifact(Git.branch, previousResultsDir, artifactName)
 
         GitHub.writeOutput("artifactName", artifactName)
         GitHub.writeOutput("artifactPath", previousResultsDir.absolutePathString())
@@ -52,69 +46,5 @@ class ResultProcessor {
 
         // Copy the latest test run results to the archived result dir.
         ResultsSet(previousResultsDir).add(latestResults.path, info = Git.hash)
-    }
-
-    private fun downloadResults() {
-        baselineResultsDir.createDirectories()
-        previousResultsDir.createDirectories()
-
-        if (!env.containsKey("GITHUB_TOKEN")) {
-            return
-        }
-
-        val github = GitHubBuilder().withOAuthToken(env["GITHUB_TOKEN"]!!).build()
-        val repo = github.getRepository(env["GITHUB_REPOSITORY"]!!)
-        val workflow = repo.listWorkflows().single {
-            // GITHUB_WORKFLOW: The name of the workflow. For example, My test workflow. If the workflow file doesn't
-            // specify a name, the value of this variable is the full path of the workflow file in the repository.
-            listOf(it.name, it.path).contains(env["GITHUB_WORKFLOW"]!!)
-        }
-        downloadResultsFor(workflow, Git.baseBranch, baselineResultsDir)
-        downloadResultsFor(workflow, Git.branch, previousResultsDir)
-    }
-
-    private fun downloadResultsFor(workflow: GHWorkflow, branch: String, targetDir: Path) {
-        println("Trying to download previous results for branch $branch")
-
-        val run = workflow.listRuns()
-            .firstOrNull { it.headBranch == branch && it.conclusion == GHWorkflowRun.Conclusion.SUCCESS }
-        if (run == null) {
-            println("Couldn't find any successful run workflow ${workflow.name}")
-            return
-        }
-
-        val artifact = run.listArtifacts().firstOrNull { it.name == artifactName }
-        if (artifact == null) {
-            println("Couldn't find any artifact matching $artifactName")
-            return
-        }
-
-        println("Downloading artifact ${artifact.archiveDownloadUrl} and extracting to $targetDir")
-        artifact.download {
-            val zipStream = ZipInputStream(it)
-            var entry: ZipEntry?
-            // while there are entries I process them
-            while (true) {
-                entry = zipStream.nextEntry
-                if (entry == null) {
-                    break
-                }
-                if (entry.isDirectory) {
-                    Path.of(entry.name).createDirectories()
-                } else {
-                    println("Extracting ${entry.name}")
-                    val outFile = FileOutputStream(targetDir.resolve(entry.name).toFile())
-                    while (zipStream.available() > 0) {
-                        val c = zipStream.read()
-                        if (c > 0) {
-                            outFile.write(c)
-                        } else {
-                            break
-                        }
-                    }
-                    outFile.close()
-                }
-            }
-        }
     }
 }
